Add React + JSX support
To use React for JSX + VirtualDom babel and Webpack needed to be added in addition to a build step.  The public directory has also been changed over to dist and is regenerated from src + root's images folder on each build.  There are issues with webpack + hmr that will be resolved at a later date and the whole application isn't within a React Component yet.