Use typescript with node, food prefs on api
Changes the build process so typescript can be used with node (all of the typescript is under node_ts and compiles to dist_node) and adds the ability to delete and adds food preference, GET+POST+DELETE to the api.  This currently doesn't write to a database, the data is lost on server restart, and the application doesn't use it yet.