add user uuid, require bearer token for delete+post, async xhr
Generate a UUID for a user, protect the routes that modify data with jwt, and while switching over to axios find a way to make the XHR asynchronous.  This did require the results to not be taken from the prop but instead from local storage since it was also being stored there.